New Design for Skills Section and improved layout
still sorting out how to get the ‘resume’ button to orient well on
small screens, fixed 7 or 8 other imperfections in the layout, changed
a bunch of text to adapt.

Went with an interesting approach, shrinking all to scale on the making
smaller of the screen size, and worked with columns on the bottom for
responsive layout functions.